+++ tests/Mcfedr/QueueManagerBundle/Tests/DependencyInjection/McfedrQueueManagerExtensionTest.php
@@ -57,8 +57,8 @@ final class McfedrQueueManagerExtensionTest extends WebTestCase
         //Backwards compatibility
         static::assertTrue($client->getContainer()->has('mcfedr_queue_manager.registry'));
 
+        // Default subscribers added
         static::assertTrue($client->getContainer()->has(DoctrineResetSubscriber::class));
-
         static::assertTrue($client->getContainer()->has(SwiftMailerSubscriber::class));
 
         // Default managers added
@@ -71,49 +71,10 @@ final class McfedrQueueManagerExtensionTest extends WebTestCase
         $client = static::createClient([
             'environment' => 'test_no_doctrine',
         ]);
-        static::assertTrue($client->getContainer()->has(TestQueueManager::class));
-        $service = $client->getContainer()->get(TestQueueManager::class);
-        static::assertInstanceOf(TestQueueManager::class, $service);
-        $options = $service->getOptions();
-        static::assertSame('127.0.0.2', $options['host']);
-        static::assertSame('mcfedr_queue', $options['default_queue']);
-        static::assertSame(1234, $options['port']);
-        static::assertSame(3, $options['retry_limit']);
-        static::assertSame(5, $options['sleep_seconds']);
-
-        //Backwards compatibility
-        static::assertTrue($client->getContainer()->has('mcfedr_queue_manager.default'));
-        $service = $client->getContainer()->get('mcfedr_queue_manager.default');
-        static::assertInstanceOf(TestQueueManager::class, $service);
-
-        $parameterOptions = $client->getContainer()->getParameter('mcfedr_queue_manager.default.options');
-        static::assertSame('127.0.0.2', $parameterOptions['host']);
-        static::assertSame('mcfedr_queue', $parameterOptions['default_queue']);
-        static::assertSame(1234, $parameterOptions['port']);
-        static::assertSame(3, $parameterOptions['retry_limit']);
-        static::assertSame(5, $parameterOptions['sleep_seconds']);
-
-        static::assertTrue($client->getContainer()->has(TestRunnerCommand::class));
-        /** @var TestRunnerCommand $command */
-        $command = $client->getContainer()->get(TestRunnerCommand::class);
-        static::assertInstanceOf(TestRunnerCommand::class, $command);
-        static::assertSame('mcfedr:queue:default-runner', $command->getName());
-        $commandOptions = $command->getOptions();
-        static::assertSame('127.0.0.2', $commandOptions['host']);
-        static::assertSame('mcfedr_queue', $commandOptions['default_queue']);
-        static::assertSame(1234, $commandOptions['port']);
-        static::assertSame(3, $commandOptions['retry_limit']);
-        static::assertSame(5, $commandOptions['sleep_seconds']);
-
-        static::assertTrue($client->getContainer()->has(QueueManagerRegistry::class));
-        //Backwards compatibility
-        static::assertTrue($client->getContainer()->has('mcfedr_queue_manager.registry'));
-
+        // Default subscribers not added
         static::assertFalse($client->getContainer()->has(DoctrineResetSubscriber::class));
-
         static::assertFalse($client->getContainer()->has(SwiftMailerSubscriber::class));
-
-        // Default managers added
+        // Default managers not added
         static::assertFalse($client->getContainer()->has('mcfedr_queue_manager.delay'));
         static::assertFalse($client->getContainer()->has('mcfedr_queue_manager.periodic'));
     }