+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 6.5.0
 
 - Add default period and delay managers.
+- Add support for getName in Worker that replaces the default name.
 
 ### 6.4.4